requirements.txt
streamlit
tensorflow
pillow
gtts